safer cleanup

+++ scripts/agent/stop.sh
@@ -5,7 +5,12 @@ sudo /etc/init.d/bixby stop
 sudo pkill -9 -u 0 -U 0 -f god
 sudo pkill -9 -u 0 -U 0 -f bixby-agent
 sudo pkill -9 -f mon_daemon
-cd /opt/bixby/etc
-sudo rm -f bixby.yml id_rsa server.pub god.d/m*.god
-cd ../var/
-sudo rm -rf *
+
+if [[ -f /opt/bixby/etc ]]; then
+  cd /opt/bixby/etc
+  sudo rm -f bixby.yml id_rsa server.pub god.d/m*.god
+fi
+
+if [[ -f /opt/bixby/var ]];
+  sudo rm -rf /opt/bixby/var/*
+fi
